requirements.txt
Flask
Flask-SQLAlchemy
psycopg2
Faker
avro-python3
avro_validator